+++ database/factories/SitesRejectedDomainFactory.php
@@ -21,14 +21,16 @@
 
 declare(strict_types=1);
 
-use Adshares\Adserver\Http\Controllers\Manager\ServerConfigurationController;
-use Adshares\Adserver\Http\Kernel;
-use Illuminate\Support\Facades\Route;
+namespace Database\Factories;
 
-Route::middleware([Kernel::ADMIN_JWT_ACCESS, Kernel::JSON_API_NO_TRANSFORM])->group(function () {
-    Route::get('config/placeholders/{key?}', [ServerConfigurationController::class, 'fetchPlaceholders']);
-    Route::patch('config/placeholders', [ServerConfigurationController::class, 'storePlaceholders']);
-    Route::get('config/{key?}', [ServerConfigurationController::class, 'fetch']);
-    Route::patch('config', [ServerConfigurationController::class, 'store']);
-    Route::put('config/{key}', [ServerConfigurationController::class, 'storeOne']);
-});
+use Illuminate\Database\Eloquent\Factories\Factory;
+
+class SitesRejectedDomainFactory extends Factory
+{
+    public function definition(): array
+    {
+        return [
+            'domain' => $this->faker->domainName(),
+        ];
+    }
+}