+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Possibility to set application currency other than ADS
+- AdPanel's placeholders to config API
 ### Changed
 - Split user access for advertisers and publishers
 ### Fixed